Reorganize tests in ddf-test Remove unused files in ddf-test Refactor the framework Switch to FunSuite
