+++ ddf-test/src/main/scala/io/ddf/test/it/etl/ReshapingHandlerSuite.scala
@@ -1,0 +1,27 @@
+/*
+ * Licensed to Tuplejump Software Pvt. Ltd. under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. Tuplejump Software Pvt. Ltd. licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except)compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to)writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ *
+ */
+package io.ddf.test.it.etl
+
+import io.ddf.test.it.BaseSuite
+import org.scalatest.Matchers
+
+trait ReshapingHandlerSuite extends BaseSuite with Matchers {
+
+  // TODO: add test cases, now this is just a placeholder
+}
